+++ src/attributes/utils/data.ts
@@ -10,10 +10,13 @@ import { FormsetData } from "@saleor/hooks/useFormset";
 import { PageDetails_page_attributes } from "@saleor/pages/types/PageDetails";
 import { ProductDetails_product_attributes } from "@saleor/products/types/ProductDetails";
 import { SearchPages_search_edges_node } from "@saleor/searches/types/SearchPages";
+import { SearchProducts_search_edges_node } from "@saleor/searches/types/SearchProducts";
 import {
+  AttributeEntityTypeEnum,
   AttributeInputTypeEnum,
   AttributeValueInput
 } from "@saleor/types/globalTypes";
+import { mapNodeToChoice, mapPagesToChoices } from "@saleor/utils/maps";
 import { MutationFetchResult } from "react-apollo";
 
 import { AttributePageFormData } from "../components/AttributePage";
@@ -24,6 +27,11 @@ export const ATTRIBUTE_TYPES_WITH_DEDICATED_VALUES = [
   AttributeInputTypeEnum.DROPDOWN,
   AttributeInputTypeEnum.MULTISELECT
 ];
+
+export interface AttributeReference {
+  label: string;
+  value: string;
+}
 
 function getSimpleAttributeData(
   data: AttributePageFormData,
@@ -199,7 +207,7 @@ export const getFileAttributeDisplayData = (
   return attribute;
 };
 
-export const getReferenceAttributeDisplayData = (
+export const getPageReferenceAttributeDisplayData = (
   attribute: AttributeInput,
   referencePages: SearchPages_search_edges_node[]
 ) => ({
@@ -208,21 +216,67 @@ export const getReferenceAttributeDisplayData = (
     ...attribute.data,
     references:
       referencePages?.length > 0 && attribute.value?.length > 0
-        ? attribute.value.map(value =>
-            referencePages.find(reference => reference.id === value)
+        ? mapPagesToChoices(
+            attribute.value.map(value => {
+              const reference = referencePages.find(
+                reference => reference.id === value
+              );
+              return { ...reference };
+            })
           )
         : []
   }
 });
 
+export const getProductReferenceAttributeDisplayData = (
+  attribute: AttributeInput,
+  referenceProducts: SearchProducts_search_edges_node[]
+) => ({
+  ...attribute,
+  data: {
+    ...attribute.data,
+    references:
+      referenceProducts?.length > 0 && attribute.value?.length > 0
+        ? mapNodeToChoice(
+            attribute.value.map(value => {
+              const reference = referenceProducts.find(
+                reference => reference.id === value
+              );
+              return { ...reference };
+            })
+          )
+        : []
+  }
+});
+
+export const getReferenceAttributeDisplayData = (
+  attribute: AttributeInput,
+  referencePages: SearchPages_search_edges_node[],
+  referenceProducts: SearchProducts_search_edges_node[]
+) => {
+  if (attribute.data.entityType === AttributeEntityTypeEnum.PAGE) {
+    return getPageReferenceAttributeDisplayData(attribute, referencePages);
+  } else if (attribute.data.entityType === AttributeEntityTypeEnum.PRODUCT) {
+    return getProductReferenceAttributeDisplayData(
+      attribute,
+      referenceProducts
+    );
+  }
+};
+
 export const getAttributesDisplayData = (
   attributes: AttributeInput[],
   attributesWithNewFileValue: FormsetData<null, File>,
-  referencePages: SearchPages_search_edges_node[]
+  referencePages: SearchPages_search_edges_node[],
+  referenceProducts: SearchProducts_search_edges_node[]
 ) =>
   attributes.map(attribute => {
     if (attribute.data.inputType === AttributeInputTypeEnum.REFERENCE) {
-      return getReferenceAttributeDisplayData(attribute, referencePages);
+      return getReferenceAttributeDisplayData(
+        attribute,
+        referencePages,
+        referenceProducts
+      );
     }
     if (attribute.data.inputType === AttributeInputTypeEnum.FILE) {
       return getFileAttributeDisplayData(attribute, attributesWithNewFileValue);
@@ -230,3 +284,34 @@ export const getAttributesDisplayData = (
 
     return attribute;
   });
+
+export const getSelectedReferencesFromAttribute = <
+  Node extends SearchPages_search_edges_node | SearchProducts_search_edges_node
+>(
+  attribute?: AttributeInput,
+  references?: Node[]
+) =>
+  references?.filter(
+    value =>
+      !attribute?.value?.some(selectedValue => selectedValue === value.id)
+  ) || [];
+
+export const getAttributeValuesFromReferences = (
+  attributeId: string,
+  attributes?: AttributeInput[],
+  referencePages?: SearchPages_search_edges_node[],
+  referenceProducts?: SearchProducts_search_edges_node[]
+) => {
+  const attribute = attributes?.find(attribute => attribute.id === attributeId);
+
+  if (attribute?.data?.entityType === AttributeEntityTypeEnum.PAGE) {
+    return mapPagesToChoices(
+      getSelectedReferencesFromAttribute(attribute, referencePages)
+    );
+  } else if (attribute?.data?.entityType === AttributeEntityTypeEnum.PRODUCT) {
+    return mapNodeToChoice(
+      getSelectedReferencesFromAttribute(attribute, referenceProducts)
+    );
+  }
+  return [];
+};