+++ src/pages/types/PageUpdate.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { PageInput, PageErrorCode, AttributeInputTypeEnum } from "./../../types/globalTypes";
+import { PageInput, PageErrorCode, AttributeInputTypeEnum, AttributeEntityTypeEnum } from "./../../types/globalTypes";
 
 // ====================================================
 // GraphQL mutation operation: PageUpdate
@@ -36,6 +36,7 @@ export interface PageUpdate_pageUpdate_page_attributes_attribute {
   slug: string | null;
   name: string | null;
   inputType: AttributeInputTypeEnum | null;
+  entityType: AttributeEntityTypeEnum | null;
   valueRequired: boolean;
   values: (PageUpdate_pageUpdate_page_attributes_attribute_values | null)[] | null;
 }
@@ -81,6 +82,7 @@ export interface PageUpdate_pageUpdate_page_pageType_attributes {
   id: string;
   name: string | null;
   inputType: AttributeInputTypeEnum | null;
+  entityType: AttributeEntityTypeEnum | null;
   valueRequired: boolean;
   values: (PageUpdate_pageUpdate_page_pageType_attributes_values | null)[] | null;
 }