+++ src/pages/utils/data.ts
@@ -1,6 +1,5 @@
 import { getSelectedAttributeValues } from "@saleor/attributes/utils/data";
 import { AttributeInput } from "@saleor/components/Attributes";
-import { SearchPages_search_edges_node } from "@saleor/searches/types/SearchPages";
 
 import {
   PageDetails_page,
@@ -12,6 +11,7 @@ export function getAttributeInputFromPage(
 ): AttributeInput[] {
   return page?.attributes.map(attribute => ({
     data: {
+      entityType: attribute.attribute.entityType,
       inputType: attribute.attribute.inputType,
       isRequired: attribute.attribute.valueRequired,
       selectedValues: attribute.values,
@@ -28,6 +28,7 @@ export function getAttributeInputFromPageType(
 ): AttributeInput[] {
   return pageType?.attributes.map(attribute => ({
     data: {
+      entityType: attribute.entityType,
       inputType: attribute.inputType,
       isRequired: attribute.valueRequired,
       values: attribute.values
@@ -37,18 +38,3 @@ export function getAttributeInputFromPageType(
     value: []
   }));
 }
-
-export const getAttributeValuesFromReferences = (
-  attributeId: string,
-  attributes: AttributeInput[],
-  referencePages: SearchPages_search_edges_node[]
-) => {
-  const attribute = attributes?.find(attribute => attribute.id === attributeId);
-
-  return (
-    referencePages?.filter(
-      value =>
-        !attribute?.value?.some(selectedValue => selectedValue === value.id)
-    ) || []
-  );
-};